Major Progress in request handling
Now the code sends the file number, file names and encoded images in base64 through a post request, printing the response body in terminal.

For now, i haven't implemented the parser for the response yet, and i send the request to an echo api.

Actually getting the data back works, proving i am sending the data correctly and receiving reponses!

also made sure to re-convert the echoed string to an image with a web tool and it looks untouched, so no corruption afaik

I also made it so the temporary folder containing pre-processed data to send gets deleted after the call, that's good

next i will move file preprocessing and the call for the request sending to a method, to make it possible to send the folder at the press of a button (and not like it is implemented now, sending every time we load one)

this actually took so long holy